Fix installation error with brew GL package
The brew package manager for macosx supports meanwhile the installation of
x11 and GL and is not longer bound to the usage of the unsupported XQuartz
package. Using the brew packages libx11 and mesa-glu the necessary include
and library files are ínstalled in /usr/local. Unfortunately this directory
is not in the search path such that the configuration of Geant4 fails.
The fix add the the missing search path such that GL installation is found
and the configuration succeeds.