Avoid possible buffer overflow in G4strstreambuf
On shutdown there's no guarantee that there's a 0 byte at the end of the string as the 0 is only added just before printing.
So in the worst case the whole buffer could be non-zero and we print memory not outside the buffer until a zero is found.
Could also just cause random garbage to be printed.