fix(processes/scoring): get phantom param from step
This commit resolves Bugzilla/Geant4 – Problem 2636.  The changes to G4EnergySplitter here are focused on ensuring that each call to SplitEnergyInVolumes() utilizes the correct phantom parameterization.  This involves the removal of the GetPhantomParam private method in favor of fetching the parameterization from the G4Step on each invocation of SplitEnergyInVolumes().  New exceptions are thrown if the G4Step does not refer to an appropriate phantom, or if the instance variable `thePhantomParam` is required but has not yet been initialized in SplitEnergyInVolumes().